pyproject.toml
[tool.poetry]
name = "trading-system"
version = "0.1.0"
description = "Event-driven machine learning trading system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
alpaca-py = "^0.13.0"
newsapi-python = "^0.2.7"
transformers = "^4.36.0"
torch = "^2.1.0"
pandas = "^2.1.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 